♻️ Refactor: components 계층 구조 변경
LiveInfo, VideoPlayer common 바깥으로
