+++ Frontend/src/components/LiveInfo/index.tsx
