+++ Frontend/src/pages/LivePage/index.tsx
@@ -1,6 +1,6 @@
-import VideoPlayer from '@components/common/VideoPlayer';
-import LiveInfo from '@components/common/LiveInfo';
-import StreamerInfo from '@components/common/LiveInfo/StreamerInfo';
+import VideoPlayer from '@/components/VideoPlayer';
+import LiveInfo from '@/components/LiveInfo';
+import StreamerInfo from '@/components/LiveInfo/StreamerInfo';
 
 export default function LivePage() {
   return (
